requirements.txt
Flask==2.3.3
google-generativeai==0.4.1
mysql-connector-python==8.3.0
python-dotenv==1.0.1
tiktoken==0.5.2
sentence-transformers==2.2.2
flask-cors==4.0.0